refactor

+++ app/routers/category.py
@@ -1,7 +1,7 @@
 from typing import List, Annotated
 
 from fastapi.routing import APIRouter
-from fastapi import HTTPException, Path, Body, Depends
+from fastapi import HTTPException, Path, Body, Depends, status
 from sqlalchemy.orm import Session
 
 from ..schemas import CategoryReponse, CategoryCreate, CategoryUpdate
@@ -15,32 +15,38 @@ router = APIRouter(
 
 @router.get('/', response_model=List[CategoryReponse])
 def get_categories(
-    session: Session = Depends(get_db),
+    session: Annotated[Session, Depends(get_db)],
 ):
     return session.query(Category).all() 
 
 
 @router.get('/{category_id}', response_model=CategoryReponse)
 def get_one_category(
-    category_id: int = Path(ge=1),
-    session: Session = Depends(get_db),
+    category_id: Annotated[int, Path(ge=1)],
+    session: Annotated[Session, Depends(get_db)],
 ):
     category = session.query(Category).get(category_id)
     
     if not category:
-        raise HTTPException(status_code=404, detail='category not found.')
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, 
+            detail='Category not found.'
+        )
     
     return category
 
 
-@router.post('/', response_model=CategoryReponse)
+@router.post('/', response_model=CategoryReponse, status_code=status.HTTP_201_CREATED)
 def create_category(
     data: CategoryCreate,
-    session: Session = Depends(get_db)
+    session: Annotated[Session, Depends(get_db)],
 ):
     existing_category = session.query(Category).filter(Category.name==data.name).first()
     if existing_category:
-        raise HTTPException(status_code=400, detail='category exists.')
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, 
+            detail="Category with this name already exists"
+        )
 
     new_category = Category(name=data.name, description=data.description)
     session.add(new_category)
@@ -50,7 +56,7 @@ def create_category(
     return new_category
 
 
-@router.put('/{category_id}', response_model=CategoryReponse)
+@router.put('/{category_id}', response_model=CategoryReponse, status_code=status.HTTP_204_NO_CONTENT)
 def update_category(
     category_id: Annotated[int, Path(ge=1)],
     data: Annotated[CategoryUpdate, Body],
@@ -59,10 +65,10 @@ def update_category(
     existing_category = session.query(Category).get(category_id)
 
     if not existing_category:
-        raise HTTPException(status_code=404, detail='category not found.')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Category not found')
 
     if session.query(Category).filter(Category.name==data.name).first():
-        raise HTTPException(status_code=400, detail='category exists.')
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail='Category with name \'Electronics\' already exists')
 
     existing_category.name = data.name if data.name else existing_category.name
     existing_category.description = data.description if data.description else existing_category.description
@@ -73,18 +79,24 @@ def update_category(
     return existing_category
 
 
-@router.delete('/{category_id}')
-def update_category(
+@router.delete('/{category_id}', status_code=status.HTTP_204_NO_CONTENT)
+def delete_category(
     category_id: Annotated[int, Path(ge=1)],
-    session: Session = Depends(get_db)
+    session: Annotated[Session, Depends(get_db)],
 ):
     existing_category = session.query(Category).get(category_id)
 
     if not existing_category:
-        raise HTTPException(status_code=404, detail='category not found.')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Category not found')
 
+    products = existing_category.products.count()
+    if products > 0:
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=f'Cannot delete category. {products} products are linked to this category')
+    
     session.delete(existing_category)
     session.commit()
 
-    return {'message': 'deleted.'}
+    return {
+        "message": "Category deleted successfully"
+    }
